requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
audioop-lts==0.2.1
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
multidict==6.6.3
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
psycopg2-binary==2.9.10
py-cord==2.6.1
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
SQLAlchemy==2.0.42
typing_extensions==4.14.1
yarl==1.20.1
